backend/pyproject.toml
[tool.poetry]
name = "nonagon-workspace"
version = "0.1.0"
description = "Nonagon monorepo workspace - Discord quest management system"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
python-dotenv = ">=1.0"
sqlalchemy = {version = ">=2.0", extras = ["asyncio"]}
asyncpg = ">=0.29"
psycopg2-binary = ">=2.9"
fastapi = ">=0.109"
uvicorn = {version = ">=0.27", extras = ["standard"]}
strawberry-graphql = {version = ">=0.220", extras = ["fastapi"]}
"discord.py" = ">=2.3"
aiohttp = ">=3.9"
python-jose = {version = ">=3.3", extras = ["cryptography"]}
httpx = ">=0.27"
dotenv = "^0.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[poetry.scripts]
run-backend = "nonagon_workspace.backend.main:run"
